+++ sample-android-lib/src/main/java/me/tatarka/retrolambda/sample/lib/Lib.java
@@ -4,7 +4,7 @@ package me.tatarka.retrolambda.sample.lib;
  * Created by evan on 3/29/15.
  */
 public class Lib {
-    public static Function getHello() {
-        return () -> "Hello, retrolambda (from lib)!";
+    public static Function getHello(){
+        return () -> "Hello, Retrolambda (from lib)!";
     }
 }