+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+#### 3.3.0
+- Fixed certain cases where android tests would fail to compile because wrong classpath was used.
+- Bumped default retrolambda version to `2.3.0`
+
+#### 3.3.0-beta4
+- Merged in all changes up to `3.2.5`
+- Supports running with other transforms (this is a bit hackey, hopefully we'll get per-variant
+transforms and it can be implemented far more cleanly).
+
+#### 3.3.0-beta3
+- Supports transform api in android plugin `1.5.0`
+
+#### 3.3.0-beta2
+Experimental support for the new transform api
+- Requires android plugin `1.4.0-beta4`
+
+http://tools.android.com/tech-docs/new-build-system/transform-api
+
+- Transform api used in library projects.
+- Prints retrolambda output to log level --info
+known issues:
+- include/exclude are ignored
+
+#### 3.3.0-beta1
+Experimental support for the new transform api
+- Requires android plugin `1.4.0-beta2`
+
+http://tools.android.com/tech-docs/new-build-system/transform-api
+
+known issues:
+- Include/exclude are ignored.
+- Does not work in library projects (falls back to the old method).
+- Prints retrolambda output to stdout instead of log level --info.
+
 #### 3.2.5
 - Bumped default retrolambda version to `2.1.0`
 - If the commandline parameters are over a certain limit, they will be written to files. This should